[BSE-1797] Support object_keys (#6604)
Adds support for the Snowflake JSON function OBJECT_KEYS which takes in a JSON column and returns a column where each row is an array of the string keys of that row. This is our first real JSON function in BodoSQL, so some extra testing infrastructure setup (and BdooSQL context logic) was necessary to add. Logic to toggle between map vs struct arrays was ported from this PR: #6398

+++ bodo/libs/array_item_arr_ext.py
@@ -1058,8 +1058,8 @@ def array_item_arr_getitem_array(arr, ind):
                 ind += len(arr)
             offsets = get_offsets(arr)
             data = get_data(arr)
-            l_start_offset = offsets[ind]
-            l_end_offset = offsets[ind + 1]
+            l_start_offset = offsets[np.int64(ind)]
+            l_end_offset = offsets[np.int64(ind) + 1]
             return data[l_start_offset:l_end_offset]
 
         return array_item_arr_getitem_impl
